backend/requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
Flask==3.0.2
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient
packaging==23.2
PyJWT==2.8.0
requests==2.31.0
SQLAlchemy==1.4.43
urllib3==2.0.4
Werkzeug==3.0.1
zipp==3.10.0